+++ closet-application/build.gradle
@@ -15,8 +15,6 @@ project(':closet-application') {
 
         implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
 
-        implementation 'software.amazon.awssdk:s3:2.31.7'
-
         implementation 'com.nimbusds:nimbus-jose-jwt:10.3'
 
         implementation 'org.springframework.boot:spring-boot-starter-security'
